[#69] refactor : activity id 사용자 분리

+++ src/main/java/depromeet/lessonfour/server/activityrecord/app/service/DeleteActivityRecordUseCase.java
@@ -2,8 +2,8 @@ package depromeet.lessonfour.server.activityrecord.app.service;
 
 import org.springframework.transaction.annotation.Transactional;
 
-import depromeet.lessonfour.server.activityrecord.app.repository.ActivityRecordRepository;
 import depromeet.lessonfour.server.activityrecord.domain.entity.ActivityRecord;
+import depromeet.lessonfour.server.activityrecord.domain.repository.ActivityRecordRepository;
 import depromeet.lessonfour.server.common.annotation.UseCase;
 import depromeet.lessonfour.server.common.api.code.ErrorCode;
 import depromeet.lessonfour.server.common.exception.ServerException;
@@ -19,8 +19,13 @@ public class DeleteActivityRecordUseCase {
   public void deleteActivityRecord(Long userId, Long activityRecordId) {
     ActivityRecord activityRecord =
         activityRecordRepository
-            .findByUserAndId(userId, activityRecordId)
+            .findById(activityRecordId)
             .orElseThrow(() -> new ServerException(ErrorCode.DATA_NOT_FOUND));
+
+    if (!activityRecord.isOwnedBy(userId)) {
+      throw new ServerException(ErrorCode.UNAUTHORIZED);
+    }
+
     activityRecord.delete();
   }
 }
